Restructure library directory structure for explainer algorithms, move object detection example script to notebook
