+++ example/gradcam_object_detection/custom_models/ssd/ssd_object_detector.py
