+++ foxai/explainer/computer_vision/algorithm/gradcam.py
@@ -13,12 +13,13 @@ from captum._utils.typing import TargetType
 from captum.attr import GuidedGradCam
 
 from foxai.array_utils import validate_result
-from foxai.explainer.computer_vision.image_classification.base_explainer import (
-    Explainer,
-)
+from foxai.explainer.base_explainer import Explainer
 from foxai.explainer.computer_vision.model_utils import (
     get_last_conv_model_layer,
     modify_modules,
+)
+from foxai.explainer.computer_vision.object_detection.base_object_detector import (
+    BaseObjectDetector,
 )
 from foxai.explainer.computer_vision.object_detection.types import ObjectDetectionOutput
 
@@ -548,3 +549,55 @@ class LayerGradCAMCVExplainer(BaseGradCAMCVExplainer):
 
         validate_result(attributions=attributions)
         return attributions
+
+
+class LayerGradCAMObjectDetection(LayerBaseGradCAM):
+    """Layer GradCAM for object detection task.
+
+    Code based on https://github.com/pooya-mohammadi/yolov5-gradcam.
+    """
+
+    def __init__(
+        self,
+        model: BaseObjectDetector,
+        target_layer: torch.nn.Module,
+    ):
+        super().__init__(target_layer=target_layer)
+        self.model = model
+
+    def forward(
+        self,
+        input_img: torch.Tensor,
+    ) -> ObjectDetectionOutput:
+        """Forward pass of GradCAM aglorithm.
+
+        Args:
+            input_img: Input image with shape of (B, C, H, W).
+
+        Returns:
+            ObjectDetectionOutput object.
+        """
+        saliency_maps: List[torch.Tensor] = []
+        _, _, height, width = input_img.size()
+        predictions, logits = self.model.forward(input_img)
+        for logit, cls in zip(logits[0], [p.class_number for p in predictions]):
+            score = logit[cls]
+            # clear gradients
+            self.model.zero_grad()
+
+            # calculate gradients
+            score.backward(retain_graph=True)
+
+            saliency_maps.append(
+                self.get_saliency_map(
+                    height=height,
+                    width=width,
+                    gradients=self.gradients,
+                    activations=self.activations,
+                )
+            )
+        return ObjectDetectionOutput(
+            saliency_maps=saliency_maps,
+            logits=logits,
+            predictions=predictions,
+        )
