+++ foxai/explainer/computer_vision/algorithm/occlusion.py
@@ -10,9 +10,7 @@ from captum._utils.typing import TargetType
 from captum.attr import Occlusion
 
 from foxai.array_utils import validate_result
-from foxai.explainer.computer_vision.image_classification.base_explainer import (
-    Explainer,
-)
+from foxai.explainer.base_explainer import Explainer
 
 
 class OcclusionCVExplainer(Explainer):
